Add files via upload
Here is a probabilistic way to try and detect vehicles. This is not a way that I think we are going with. However, you might read the first three sections of this paper to see how they don't look at just each space individually, but at three spaces at a time.